+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ----
 ## Unreleased
 
+----
+## v1.1.2 (2018-03-07)
+
 * Added a program for plotting magnetic field lines
 * Added an option to footpoint() and farpoint() functions to normalise the output coordinates to the light cylinder radius.
 * Added an option to footpoint() and farpoint() functions to stop if the point gets too far out.